requirements.txt
certifi==2025.8.3
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
ecm-hanwha==0.1.0
filelock==3.19.1
fonttools==4.59.2
fsspec==2025.9.0
imageio==2.37.0
importlib-resources==6.5.2
jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.7
lazy-loader==0.4
markupsafe==3.0.2
matplotlib==3.9.4
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
opencv-python==4.11.0.86
packaging==25.0
pandas==2.3.2
pillow==11.3.0
psutil==7.0.0
pyparsing==3.2.3
pyproj==3.6.1
python-dateutil==2.9.0.post0
pytz==2025.2
pyxtf==1.4.2
pyyaml==6.0.2
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
six==1.17.0
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2024.8.30
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tqdm==4.67.1
typing-extensions==4.15.0
tzdata==2025.2
zipp==3.23.0